Preserve original stacktrace from sandboxed errors
The current behaviour just preserves the message from the original error, but
loses the stack so it can be hard to debug. This change adds the error's stack
onto the error object so that you have a lot more information to work with. If
no stack is present (i.e. if we're failing with a string and not an Error) then
the behaviour doesn't change.